refactor: improve and update segmentation logic and tests

+++ internal/domain/models/model.go
@@ -3,5 +3,5 @@ package models
 type UserSegment struct {
 	UserID      string `json:"user_id" validate:"required"`
 	Segment     string `json:"segment" validate:"required"`
-	RegistredAt int64
+	RegistredAt uint32
 }
